refactorizando

+++ src/main/java/es/mdef/repositorios/UsuarioDAO.java
@@ -5,6 +5,6 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import es.mdef.usuarios.Usuario;
 
 @RepositoryRestResource
-public interface UsuarioDAO extends JpaRepository<Usuario, Long> {
+public interface UsuarioDAO extends JpaRepository<Usuario, Integer> {
 
 }
